+++ Alpha/Strategies/LastClose.cpp
