+++ CMakeLists.txt
@@ -22,8 +22,8 @@ if(${CMAKE_BUILD_TYPE} STREQUAL "Debug")
 endif()
 
 set(Platform_Inc "${CMAKE_CURRENT_SOURCE_DIR}/Platform/inc/")
-set(Indicators_Inc "${CMAKE_CURRENT_SOURCE_DIR}/Alpha/Indicators/inc/")
-set(Strategies_Inc "${CMAKE_CURRENT_SOURCE_DIR}/Alpha/Strategies/inc/")
+#set(Indicators_Inc "${CMAKE_CURRENT_SOURCE_DIR}/Alpha/Indicators/inc/")
+#set(Strategies_Inc "${CMAKE_CURRENT_SOURCE_DIR}/Alpha/Strategies/inc/")
 
 find_package(spdlog CONFIG REQUIRED)
 
@@ -51,8 +51,7 @@ if(${ENABLE_LINTER})
     endif()
 endif()
 
-add_subdirectory("Alpha")
-add_subdirectory("Backtesting")
+add_subdirectory("Test")
 add_subdirectory("Platform")
 
 set(CLANG_FORMAT_EXCLUDE_PATTERNS "build" "vcpkg")
