+++ Platform/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB SOURCES "*.cpp" )
+file(GLOB SOURCES "*.cpp")
 add_library(TBPlatform SHARED ${SOURCES})
 set_target_properties(TBPlatform
     PROPERTIES
@@ -6,7 +6,7 @@ set_target_properties(TBPlatform
     LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib" 
 )
 target_compile_options(TBPlatform PRIVATE -fPIC)
-target_include_directories(TBPlatform PRIVATE "inc/" ${Indicators_Inc} ${Strategies_Inc})
+target_include_directories(TBPlatform PRIVATE "inc/")
 target_link_libraries(TBPlatform PRIVATE spdlog::spdlog spdlog::spdlog_header_only)
 install(TARGETS TBPlatform
     LIBRARY DESTINATION lib
