+++ Platform/inc/Util.hpp
@@ -1,7 +1,6 @@
 #pragma once
-#include "Account.h"
+#include "BackTrader.h"
 #include "Brain.h"
-#include "Broker.h"
 #include <iomanip>
 #include <iostream>
 #include <memory>
@@ -18,18 +17,18 @@ struct InputFile
     std::string exchange;
     std::string currency;
     std::string exprDate;
-    bool candles;
-    bool snaps;
-    bool options;
+    bool        candles;
+    bool        snaps;
+    bool        options;
     /// Candle or Snap constructor
     InputFile( std::string path, std::string newSym, std::string newSecId, std::string newSecType, std::string newEx, std::string newCur, bool candle = true )
     {
-        filepath = std::move(path);
-        symbol = std::move(newSym);
-        secId = std::move(newSecId);
-        secType = std::move(newSecType);
-        exchange = std::move(newEx);
-        currency = std::move(newCur);
+        filepath = std::move( path );
+        symbol = std::move( newSym );
+        secId = std::move( newSecId );
+        secType = std::move( newSecType );
+        exchange = std::move( newEx );
+        currency = std::move( newCur );
         exprDate = "";
         candles = candle;
         snaps = !candle;
@@ -51,21 +50,40 @@ struct InputFile
     }
 };
 
+/// Helpful for shrinking strategy class codes.
+inline Order getOrder( bool buy, double quant = 1, std::string timeForce = "DAY", std::string oT = "MKT" )
+{
+    auto newOrder = Order();
+    newOrder.orderId = BackTrader::getNextOrderId();
+    newOrder.totalQuantity = quant;
+    newOrder.tif = timeForce;
+    newOrder.orderType = oT;
+    if( buy )
+    {
+        newOrder.action = "BUY";
+    }
+    else
+    {
+        newOrder.action = "SELL";
+    }
+    return newOrder;
+}
+
 /// @brief Parses positional input args into filepaths
 ///
 /// Use this method if your input data file already has its information at the top.
 /// This method must be used if your file does not have its data column headers on the first line of the file.
 inline std::vector<std::string> ReadInputs( int argc, char* argv[] )
 {
-    if( (argc-1) == 0 )
+    if( ( argc - 1 ) == 0 )
     {
         std::cout << "You must provide an input file path as a positional argument!" << std::endl;
-        exit(-1);
+        exit( -1 );
     }
     auto retPaths = std::vector<std::string>();
     for( int i = 1; i < argc; i++ )
     {
-        retPaths.push_back(argv[i]);
+        retPaths.push_back( argv[i] );
     }
     return retPaths;
 }
@@ -75,10 +93,10 @@ inline std::vector<std::string> ReadInputs( int argc, char* argv[] )
 /// tickerInfos should contain the information about an input csv, but not the filepath.
 inline void ReadInputs( int argc, char* argv[], InputFile& tickerInfo )
 {
-    if( (argc-1) == 0 )
+    if( ( argc - 1 ) == 0 )
     {
         std::cout << "You must provide an input file path as a positional argument!" << std::endl;
-        exit(-1);
+        exit( -1 );
     }
     tickerInfo.filepath = argv[1];
 }
@@ -88,33 +106,33 @@ inline void ReadInputs( int argc, char* argv[], InputFile& tickerInfo )
 /// tickerInfos should contain the information about an input csv, but not the filepath.
 inline void ReadInputs( int argc, char* argv[], std::vector<InputFile>& tickerInfos )
 {
-    if( (argc-1) != tickerInfos.size() )
+    if( ( argc - 1 ) != tickerInfos.size() )
     {
         std::cout << "Input arguments and file information lengths do not match!" << std::endl;
-        exit(-1);
+        exit( -1 );
     }
     for( int i = 1; i < argc; i++ )
     {
-        tickerInfos[i-1].filepath = argv[i];
+        tickerInfos[i - 1].filepath = argv[i];
     }
 }
 
 inline void PrintResults( const std::unique_ptr<BTBrain>& brain )
 {
     std::cout << std::fixed << std::setprecision( 2 );
-    std::cout << "Ending cash was " << brain->Account->getCash() << std::endl;
-    std::cout << "Ending margin loan was " << brain->Account->getMarginLoan() << std::endl;
-    std::cout << "Total commission was " << brain->Broker->getTotalCommission() << std::endl;
-    std::cout << "PnL was " << brain->Account->getPnL() << std::endl;
-    std::cout << "UPnL was " << brain->Account->getUPnL() << std::endl;
-    std::cout << "Total account change was " << brain->Account->getUPnL()+brain->Account->getPnL() << std::endl;
+    std::cout << "Ending cash was " << brain->getCash() << std::endl;
+    std::cout << "Ending margin loan was " << brain->getMarginLoan() << std::endl;
+    std::cout << "Total commission was " << brain->getTotalCommission() << std::endl;
+    std::cout << "PnL was " << brain->getPnL() << std::endl;
+    std::cout << "UPnL was " << brain->getUPnL() << std::endl;
+    std::cout << "Total account change was " << brain->getUPnL() + brain->getPnL() << std::endl;
     std::cout << "Ending positions are " << std::endl;
-    for( auto& index : brain->Account->positions )
+    for( auto& index : brain->getPositions() )
     {
         std::string name = index->getContract()->symbol;
-        if( index->getContract()->secType == "OPT")
+        if( index->getContract()->secType == "OPT" )
         {
-            name += std::to_string(index->getContract()->strike);
+            name += std::to_string( index->getContract()->strike );
             name += index->getContract()->right;
         }
         std::cout << index->getContract()->symbol << ", " << index->getAvgPrice() << ", " << index->getPositionSize() << std::endl;
